WIP: Extracted public superclass of CliFunctionResult
