+++ geode-core/src/main/java/org/apache/geode/management/internal/cli/functions/CliFunctionResult.java
@@ -18,27 +18,14 @@ import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 
 import org.apache.geode.DataSerializer;
-import org.apache.geode.internal.DataSerializableFixedID;
-import org.apache.geode.internal.Version;
+import org.apache.geode.management.FunctionResult;
 import org.apache.geode.management.internal.configuration.domain.XmlEntity;
 
-public class CliFunctionResult implements Comparable<CliFunctionResult>, DataSerializableFixedID {
-  private String memberIdOrName;
-  private Serializable[] serializables = new String[0];
-  private Object resultObject;
+public class CliFunctionResult extends FunctionResult {
   private XmlEntity xmlEntity;
-  private byte[] byteData = new byte[0];
-  private StatusState state;
-
-  public enum StatusState {
-    OK, ERROR, IGNORABLE
-  }
 
   @Deprecated
   public CliFunctionResult() {}
@@ -118,50 +105,12 @@ public class CliFunctionResult implements Comparable<CliFunctionResult>, DataSer
     this(memberIdOrName, resultObject, null);
   }
 
-  public String getMemberIdOrName() {
-    return this.memberIdOrName;
-  }
-
-  @Deprecated
-  public String getMessage() {
-    if (this.serializables.length == 0 || !(this.serializables[0] instanceof String)) {
-      return null;
-    }
-
-    return (String) this.serializables[0];
-  }
-
   public String getStatus(boolean skipIgnore) {
     if (state == StatusState.IGNORABLE) {
       return skipIgnore ? "IGNORED" : "ERROR";
     }
 
     return state.name();
-  }
-
-  public String getStatus() {
-    return getStatus(true);
-  }
-
-  public String getStatusMessage() {
-    String message = getMessage();
-
-    if (isSuccessful()) {
-      return message;
-    }
-
-    String errorMessage = "";
-    if (message != null
-        && (resultObject == null || !((Throwable) resultObject).getMessage().contains(message))) {
-      errorMessage = message;
-    }
-
-    if (resultObject != null) {
-      errorMessage = errorMessage.trim() + " " + ((Throwable) resultObject).getClass().getName()
-          + ": " + ((Throwable) resultObject).getMessage();
-    }
-
-    return errorMessage;
   }
 
   /**
@@ -187,28 +136,6 @@ public class CliFunctionResult implements Comparable<CliFunctionResult>, DataSer
     }
 
     return errorMessage;
-  }
-
-  @Deprecated
-  public Serializable[] getSerializables() {
-    return this.serializables;
-  }
-
-  @Deprecated
-  public Throwable getThrowable() {
-    if (isSuccessful()) {
-      return null;
-    }
-    return ((Throwable) resultObject);
-  }
-
-  public Object getResultObject() {
-    return resultObject;
-  }
-
-  @Override
-  public int getDSFID() {
-    return DataSerializableFixedID.CLI_FUNCTION_RESULT;
   }
 
   @Override
@@ -253,10 +180,6 @@ public class CliFunctionResult implements Comparable<CliFunctionResult>, DataSer
     this.serializables = (Serializable[]) DataSerializer.readObjectArray(in);
   }
 
-  public boolean isSuccessful() {
-    return this.state == StatusState.OK;
-  }
-
   public boolean isIgnorableFailure() {
     return this.state == StatusState.IGNORABLE;
   }
@@ -266,79 +189,11 @@ public class CliFunctionResult implements Comparable<CliFunctionResult>, DataSer
     return this.xmlEntity;
   }
 
-  @Deprecated
-  public byte[] getByteData() {
-    return this.byteData;
-  }
-
-  @Override
-  public int compareTo(CliFunctionResult o) {
-    if (this.memberIdOrName == null && o.memberIdOrName == null) {
-      return 0;
-    }
-    if (this.memberIdOrName == null && o.memberIdOrName != null) {
-      return -1;
-    }
-    if (this.memberIdOrName != null && o.memberIdOrName == null) {
-      return 1;
-    }
-    return getMemberIdOrName().compareTo(o.memberIdOrName);
-  }
-
-  @Override
-  public int hashCode() {
-    final int prime = 31;
-    int result = 1;
-    result = prime * result + ((this.memberIdOrName == null) ? 0 : this.memberIdOrName.hashCode());
-    return result;
-  }
-
-  @Override
-  public boolean equals(Object obj) {
-    if (this == obj)
-      return true;
-    if (obj == null)
-      return false;
-    if (getClass() != obj.getClass())
-      return false;
-    CliFunctionResult other = (CliFunctionResult) obj;
-    if (this.memberIdOrName == null) {
-      if (other.memberIdOrName != null)
-        return false;
-    } else if (!this.memberIdOrName.equals(other.memberIdOrName))
-      return false;
-    return true;
-  }
-
   @Override
   public String toString() {
     return "CliFunctionResult [memberId=" + this.memberIdOrName + ", successful="
         + this.isSuccessful() + ", xmlEntity=" + this.xmlEntity + ", serializables="
         + Arrays.toString(this.serializables) + ", throwable=" + this.resultObject + ", byteData="
         + Arrays.toString(this.byteData) + "]";
-  }
-
-  /**
-   * Remove elements from the list that are not instances of CliFunctionResult and then sort the
-   * results.
-   *
-   * @param results The results to clean.
-   * @return The cleaned results.
-   */
-  public static List<CliFunctionResult> cleanResults(List<?> results) {
-    List<CliFunctionResult> returnResults = new ArrayList<CliFunctionResult>(results.size());
-    for (Object result : results) {
-      if (result instanceof CliFunctionResult) {
-        returnResults.add((CliFunctionResult) result);
-      }
-    }
-
-    Collections.sort(returnResults);
-    return returnResults;
-  }
-
-  @Override
-  public Version[] getSerializationVersions() {
-    return new Version[] {Version.GFE_80};
   }
 }
