+++ geode-core/src/test/java/org/apache/geode/management/internal/cli/result/model/ResultModelTest.java
@@ -15,9 +15,9 @@
 
 package org.apache.geode.management.internal.cli.result.model;
 
-import static org.apache.geode.management.internal.cli.functions.CliFunctionResult.StatusState.ERROR;
-import static org.apache.geode.management.internal.cli.functions.CliFunctionResult.StatusState.IGNORABLE;
-import static org.apache.geode.management.internal.cli.functions.CliFunctionResult.StatusState.OK;
+import static org.apache.geode.management.FunctionResult.StatusState.ERROR;
+import static org.apache.geode.management.FunctionResult.StatusState.IGNORABLE;
+import static org.apache.geode.management.FunctionResult.StatusState.OK;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
